backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
alembic==1.17.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-multipart==0.0.22
python-dotenv==1.2.2
werkzeug==3.1.6
cloudinary==1.44.1
